Generate parameter's exmaple under parameter itself, not under its schema
