+++ src/Swashbuckle.AspNetCore.SwaggerGen/XmlComments/XmlCommentsParameterFilter.cs
@@ -1,5 +1,9 @@
-﻿using System.Reflection;
+﻿using System;
+using System.ComponentModel;
+using System.Globalization;
+using System.Reflection;
 using System.Xml.XPath;
+using Microsoft.OpenApi.Any;
 using Microsoft.OpenApi.Models;
 
 namespace Swashbuckle.AspNetCore.SwaggerGen
@@ -52,7 +56,32 @@ namespace Swashbuckle.AspNetCore.SwaggerGen
             if (paramNode != null)
             {
                 parameter.Description = XmlCommentsTextHelper.Humanize(paramNode.InnerXml);
+
+                var example = paramNode.GetAttribute("example", string.Empty);
+                if (!string.IsNullOrEmpty(example))
+                {
+                    parameter.Example = ConvertToOpenApiType(parameterInfo.ParameterType, parameter.Schema, example);
+                }
             }
+        }
+
+        private static IOpenApiAny ConvertToOpenApiType(Type memberType, OpenApiSchema schema, string stringValue)
+        {
+            object typedValue;
+
+            try
+            {
+                typedValue = TypeDescriptor.GetConverter(memberType).ConvertFrom(
+                    context: null,
+                    culture: CultureInfo.InvariantCulture,
+                    stringValue);
+            }
+            catch (Exception)
+            {
+                return null;
+            }
+
+            return OpenApiAnyFactory.CreateFor(schema, typedValue);
         }
     }
 }
